requirements.txt
beautifulsoup4==4.13.4
cached-property==2.0.1
certifi==2025.6.15
charset-normalizer==3.4.2
chevron==0.14.0
deepl==1.22.0
frozendict==2.4.6
genanki==0.13.1
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.19.1
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
soupsieve==2.7
typing_extensions==4.14.0
urllib3==2.5.0
